requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
cfgv==3.5.0
charset-normalizer==3.4.4
distlib==0.4.0
Django==4.2
django-cors-headers==3.13.0
django-debug-toolbar==4.0.0
django-extra-fields==3.0.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.7.0
drf-spectacular==0.27.2
filelock==3.20.0
filetype==1.2.0
flake8==7.3.0
identify==2.6.15
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
nodeenv==1.9.1
pillow==10.4.0
platformdirs==4.5.0
pre-commit==3.6.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.37.0
requests==2.31.0
rpds-py==0.29.0
setuptools==80.9.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.35.4
